netdisco/requirements.txt
Package            Version
------------------ ---------
certifi            2021.10.8
charset-normalizer 2.0.12
idna               3.3
numpy              1.22.3
pandas             1.4.1
pip                22.0.4
python-dateutil    2.8.2
python-dotenv      0.20.0
pytz               2022.1
requests           2.27.1
setuptools         58.1.0
six                1.16.0
urllib3            1.26.9
